Moved the file validations.js to a users folder to emphasize that the validations are refering to users
